requirements.txt
# ─── Core Data Science ──────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1

# ─── Machine Learning & Survival Analysis ───────────────────────────────────────
scikit-learn==1.5.0
lifelines==0.29.0

# ─── NLP & Transformers ─────────────────────────────────────────────────────────
transformers==4.41.2
torch==2.3.1                    # CPU-only: torch==2.3.1+cpu (ver nota al pie)
tokenizers==0.19.1
feedparser==6.0.11              # Parseo de feeds RSS/Atom

# ─── HTTP & Web Scraping ────────────────────────────────────────────────────────
requests==2.32.3
urllib3==2.2.1
beautifulsoup4==4.12.3
lxml==5.2.2

# ─── Geospatial & Visualization ─────────────────────────────────────────────────
folium==0.17.0
geopandas==0.14.4
shapely==2.0.4
branca==0.7.2

# ─── Configuration & Environment ───────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.4                 # Validación de schemas de datos

# ─── Logging & Monitoring ───────────────────────────────────────────────────────
structlog==24.2.0               # Logging estructurado JSON-compatible

# ─── Testing & Synthetic Data ───────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0
faker==25.2.0               # Generación de datos sintéticos realistas

# ─── Notebooks ──────────────────────────────────────────────────────────────────
jupyter==1.0.0
ipykernel==6.29.4
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# ─── Code Quality ───────────────────────────────────────────────────────────────
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.10.0
pre-commit==3.7.1

# ─── Data Versioning (opcional, recomendado) ────────────────────────────────────
# dvc==3.51.2

# ─────────────────────────────────────────────────────────────────────────────────
# NOTA: Para instalación CPU-only de PyTorch en Windows:
#   pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu
# Para GPU (CUDA 12.1):
#   pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cu121
